refactor: apply AI code review feedback (#10)
