+++ docs/ai-review.md
@@ -1,0 +1,54 @@
+# Step 5. AI 코드리뷰 및 리팩토링
+
+## 1. 리뷰 전 코드의 문제점
+
+- PostService가 게시글 생성뿐 아니라 사용자 저장 책임까지 함께 가지고 있어
+  단일 책임 원칙(SRP)을 위반하고 있었다.
+- Controller에서 테스트를 위한 임시 User 도메인 객체를 직접 생성하여
+  프로덕션 코드와 테스트 코드의 경계가 모호했다.
+- 생성 시간(createdAt)을 외부(Service)에서 주입하고 있어
+  도메인 책임이 분산되어 있었다.
+
+---
+
+## 2. AI 코드리뷰 주요 피드백
+
+### (1) SOLID 원칙
+- Service는 하나의 비즈니스 책임만 가져야 하며,
+  PostService는 "게시글 생성"에만 집중하는 것이 바람직하다는 피드백을 받았다.
+- Controller는 HTTP 요청/응답 처리에만 집중하고
+  도메인 객체 생성 책임을 최소화해야 한다는 점을 인지했다.
+
+### (2) 이펙티브 자바 관점
+- 생성자를 통해 필수 필드를 강제한 점은 긍정적이나,
+  도메인 내부에서 책임질 수 있는 속성은
+  엔티티 스스로 관리하는 것이 더 명확하다는 피드백을 받았다.
+
+### (3) 아키텍처 관점
+- 레이어드 아키텍처(Controller → Service → Repository)는 잘 유지되고 있었으나,
+  테스트를 위한 임시 로직이 프로덕션 코드에 섞여 있어
+  가독성과 유지보수성을 저해할 수 있다는 지적이 있었다.
+
+---
+
+## 3. 리팩토링 내용
+
+- PostService에서 UserRepository 의존성을 제거하고
+  게시글 생성 책임만 남기도록 수정하였다.
+- Controller에서 임시 사용자 생성 로직을 별도 메서드로 분리하여
+  테스트용 코드임을 명확히 드러냈다.
+- Post 엔티티가 생성 시간을 스스로 책임지도록 변경하여
+  도메인 응집도를 높였다.
+- 게시글 생성 요청 DTO를 별도 클래스로 분리하여
+  Controller 코드의 역할을 명확히 했다.
+
+---
+
+## 4. 리팩토링 후 느낀 점
+
+- 각 레이어의 책임이 명확해지면서 코드의 가독성이 크게 향상되었다.
+- 단순히 동작하는 코드를 작성하는 것과
+  역할과 책임을 고려해 설계하는 것은 큰 차이가 있다는 점을 체감했다.
+- AI 코드리뷰를 통해
+  스스로 인지하지 못했던 설계상의 문제를 발견하고
+  개선하는 경험을 할 수 있었다.
