+++ src/main/java/com/cmc/board/post/controller/PostController.java
@@ -8,9 +8,6 @@ import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
 
-/**
- * 게시글 API Controller
- */
 @RestController
 @RequestMapping("/posts")
 public class PostController {
@@ -21,25 +18,13 @@ public class PostController {
         this.postService = postService;
     }
 
-    /**
-     * 게시글 생성 API (임시 버전)
-     *
-     * - 로그인/세션 구현 전이라
-     *   User를 임의로 생성해서 사용
-     */
     @PostMapping
     public ResponseEntity<Post> createPost(@RequestBody CreatePostRequest request) {
 
-        // ⚠️ 임시 사용자 (나중에 세션에서 꺼낼 예정)
-        User tempUser = new User(
-                "test@test.com",
-                "password",
-                "USER",
-                LocalDateTime.now()
-        );
+        User author = getTemporaryUserForTest();
 
         Post post = postService.create(
-                tempUser,
+                author,
                 request.getTitle(),
                 request.getContent()
         );
@@ -48,20 +33,17 @@ public class PostController {
     }
 
     /**
-     * 게시글 생성 요청 DTO
-     * - Step 3에서는 Controller 내부 static 클래스로 둬도 OK
+     * 임시 사용자 반환 메서드
+     *
+     * - Step 3~5 단계에서 테스트를 위해 사용
+     * - 추후 세션/인증 로직으로 대체 예정
      */
-    static class CreatePostRequest {
-
-        private String title;
-        private String content;
-
-        public String getTitle() {
-            return title;
-        }
-
-        public String getContent() {
-            return content;
-        }
+    private User getTemporaryUserForTest() {
+        return new User(
+                "test@test.com",
+                "password",
+                "USER",
+                LocalDateTime.now()
+        );
     }
 }
