+++ src/main/java/com/cmc/board/post/controller/dto/CreatePostRequest.java
@@ -1,0 +1,15 @@
+package com.cmc.board.post.controller.dto;
+
+public class CreatePostRequest {
+
+    private String title;
+    private String content;
+
+    public String getTitle() {
+        return title;
+    }
+
+    public String getContent() {
+        return content;
+    }
+}
